spake2: refactor internals
Splits the crate apart into multiple modules
